\name{mrvcr-package}
\alias{mrvcr-package}
\alias{mrvcr}
\docType{package}
\title{
\packageTitle{mrvcr}
}
\description{
\packageDescription{mrvcr}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{mrvcr}

Maintainer: \packageMaintainer{mrvcr}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = mrvcr)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{mrvcr}}
%\section{Documentation Index}{\packageIndices{mrvcr}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
