% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcm.R
\name{vcm_zzt}
\alias{vcm_zzt}
\title{Variance Component Model with variance kernels V_i = Z_i \%*\% Z_i^T}
\usage{
vcm_zzt(
  formula,
  random_formula,
  data = parent.frame(),
  pet = "REML",
  method = "MM"
)
}
\description{
Variance Component Model with variance kernels V_i = Z_i \%*\% Z_i^T
}
